Extract common collapsible logic into a component
