+++ ui/src/common/index.js
@@ -2,6 +2,7 @@ export * from './AddToPlaylistButton'
 export * from './ArtistLinkField'
 export * from './BatchPlayButton'
 export * from './BitrateField'
+export * from './CollapsibleComment'
 export * from './ContextMenus'
 export * from './DateField'
 export * from './DocLink'
