+++ ui/src/themes/ligera.js
@@ -300,6 +300,8 @@ export default {
         fontSize: '.875rem',
         color: 'rgb(113 113 113 / 80%)',
       },
+    },
+    NDCollapsibleComment: {
       commentBlock: {
         fontSize: '.875rem',
         color: 'rgb(113 113 113 / 80%)',
